# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank management system
  description: api for managing user's cards
  contact:
    name: Panasik Uladzislau
    url: https://github.com/Vladislav3421730
    email: panasikvladislav1@gmail.com
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/block/{id}/status:
    put:
      tags:
        - Block Requests Management
      summary: Update block request status
      description: Updates the status of an existing block request (e.g., COMPLETED). If COMPLETED, card status will be set to BLOCKED.
      operationId: changeStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockCardRequestDto'
        required: true
      responses:
        '200':
          description: Status was changed successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BlockRequestDto'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Registration
      description: Registers a user and returns the registered user
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequestDto'
        required: true
      responses:
        '201':
          description: Successfully registered user
        '400':
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates user and returns a Access
      operationId: createToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequestDto'
        required: true
      responses:
        '200':
          description: Successfully authenticated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtResponseDto'
        '400':
          description: Invalid login credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
...
  /api/v1/cards/{id}/status:
    patch:
      tags:
        - Card Management
      summary: Update card status
      description: Changes the status of a card (e.g., ACTIVE, BLOCKED)
      operationId: changeStatus_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusCardRequestDto'
        required: true
      responses:
        '200':
          description: Card status successfully updated
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/block/{cardId}:
    patch:
      tags:
        - Block Requests Management
      summary: Create a block request for a card
      description: Creates a block request for the specified card ID if not blocked yet and if 7 days passed since last request
      operationId: createBlockRequest
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Block request was created successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/transaction/{userId}/user:
    get:
      tags:
        - Transaction Management
      summary: Get transactions by user ID
      description: Retrieve transactions associated with a specific user ID
      operationId: findAllByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Transactions not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transaction/{cardId}:
    get:
      tags:
        - Transaction Management
      summary: Get transactions by card ID
      description: Retrieve transactions associated with a specific card ID
      operationId: findAllByCardId
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Successfully find transaction
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Transactions not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/block:
    get:
      tags:
        - Block Requests Management
      summary: Get all block requests with pagination
      description: Returns a paginated list of all block requests
      operationId: findAll_3
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Requests were found successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/block/{userId}/users:
    get:
      tags:
        - Block Requests Management
      summary: Get block requests by user ID with pagination
      description: Returns a paginated list of block requests filtered by the specified user ID
      operationId: findAllByUserId_1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Requests were found successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/block/{cardId}/cards:
    get:
      tags:
        - Block Requests Management
      summary: Get block requests by card ID with pagination
      description: Returns a paginated list of block requests filtered by the specified card ID
      operationId: findAllByCardId_1
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: Requests were found successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
components:
  schemas:
...
    AppErrorDto:
      type: object
      description: DTO for errors
      properties:
        code:
          type: integer
          format: int32
          description: Error's code
          example: 400
        timestamp:
          type: string
          description: Timestamp of when the error occurred
          example: '2025-02-07 14:30:00'
        message:
          type: string
          description: Error message describing the problem
          example: Invalid request
    BlockCardRequestDto:
      type: object
      description: Request to block a card with a specific status (COMPLETED or REJECTED)
      properties:
        status:
          type: string
          description: The status of the block request. Only COMPLETED or REJECTED are allowed.
          example: COMPLETED
      required:
        - status
    BlockRequestDto:
      type: object
      description: DTO representing a block request for a card, containing status, timestamps, and related IDs.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the block request
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        createdAt:
          type: string
          format: date-time
          description: Date and time when the block request was created
          example: '2025-05-21 15:30:00'
        updatedAt:
          type: string
          format: date-time
          description: Date and time when the block request was last updated
          example: '2025-05-21 16:00:00'
        status:
          type: string
          description: Status of the block request
          enum:
            - COMPLETED
            - REJECTED
          example: COMPLETED
        cardId:
          type: string
          format: uuid
          description: Unique identifier of the card where the transaction was initiated
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
...
    RegisterUserRequestDto:
      type: object
      description: DTO for registering a new user
      properties:
        firstName:
          type: string
          description: LastName for the new user. Must be at least 3 characters long.
          example: john_doe
        lastName:
          type: string
          description: LastName for the new user. Must be at least 3 characters long.
          example: john_doe
        username:
          type: string
          description: Username of the new user. Must be a valid email format.
          example: john.doe@example.com
          maxLength: 2147483647
          minLength: 6
        password:
          type: string
          description: Password for the new user. Must be at least 6 characters long.
          example: strongpassword123
          maxLength: 2147483647
          minLength: 6
        confirmPassword:
          type: string
          description: Confirm password to verify the password entered by the user.
          example: strongpassword123
      required:
        - confirmPassword
        - firstName
        - lastName
        - password
        - username
    LoginUserRequestDto:
      type: object
      description: DTO for user login containing email and password
      properties:
        username:
          type: string
          description: User's login
          example: vlad123
        password:
          type: string
          description: User's password
          example: q1w2e3
      required:
        - password
        - username
    JwtResponseDto:
      type: object
      description: DTO for JWT response containing access and refresh tokens
      properties:
        accessToken:
          type: string
          description: Access token used for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjI4MjM2NzQ1fQ.k3rUfj1DikFbd1k6B3hJHfD9_xD6wM8uP9YzJ8gPrFs
...
    StatusCardRequestDto:
      type: object
      description: DTO for updating the card status
      properties:
        status:
          type: string
          description: New card status. The allowed values are defined in the custom @Status annotation
          example: ACTIVE
      required:
        - status
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          $ref: '#/components/schemas/PageMetadata'
...
  securitySchemes:
    BearerAuthentication:
      type: http
      scheme: bearer
      bearerFormat: JWT